♻️ Move iscn badge to main setting
